actualización de files

+++ src/main/java/com/portfolio/mk/Repository/IPersonaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IPersonaRepository extends JpaRepository<Persona, Long> {
+public interface IPersonaRepository extends JpaRepository<Persona, Integer> {
     public Optional<Persona> findByNombre(String nombre);
     public boolean existsByNombre(String nombre);
 }
